Move shared config to utils folder, instead of importing from components folder
