+++ apps/back-office/src/app/_utils/richTextExtensions.ts
